refactor(research): move the research-ai assistant into a pro edition slot
ResearchAssistant and its exclusive children move to the pro edition. ResearchPage
now renders a three-state gate: absent shows the related-materials browse card,
locked adds LockedAiNotice, and active resolves the research-ai.assistant pro slot,
passing the browse card in as a rendered prop so the card and researchModel stay
public and shared. Extend the icon barrel with the icons the slot re-exports so the
private components import lucide through apps/web, matching the symbol-follow slot.

The old active-state render test is superseded by a narrow public gate test asserting
the three branches and the props handed to the resolved slot.
