+++ apps/web/src/ui/icons.ts
@@ -1,1 +1,17 @@
-export { RadioTower } from 'lucide-react';
+export {
+  BrainCircuit,
+  Check,
+  ChevronRight,
+  CircleAlert,
+  Database,
+  ExternalLink,
+  FileDiff,
+  FileSearch,
+  History,
+  Newspaper,
+  RadioTower,
+  RefreshCw,
+  Square,
+  Undo2,
+  X,
+} from 'lucide-react';
